+++ rem-iot/.vscode/launch.json
@@ -1,0 +1,44 @@
+// AUTOMATICALLY GENERATED FILE. PLEASE DO NOT MODIFY IT MANUALLY
+//
+// PlatformIO Debugging Solution
+//
+// Documentation: https://docs.platformio.org/en/latest/plus/debugging.html
+// Configuration: https://docs.platformio.org/en/latest/projectconf/sections/env/options/debug/index.html
+
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug",
+            "executable": "/Users/kevinlutzer/Projects/room-environment-monitor/rem-iot/.pio/build/esp32-c3-devkitm-1/firmware.elf",
+            "projectEnvName": "esp32-c3-devkitm-1",
+            "toolchainBinDir": "/Users/kevinlutzer/.platformio/packages/toolchain-riscv32-esp/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "preLaunchTask": {
+                "type": "PlatformIO",
+                "task": "Pre-Debug"
+            }
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (skip Pre-Debug)",
+            "executable": "/Users/kevinlutzer/Projects/room-environment-monitor/rem-iot/.pio/build/esp32-c3-devkitm-1/firmware.elf",
+            "projectEnvName": "esp32-c3-devkitm-1",
+            "toolchainBinDir": "/Users/kevinlutzer/.platformio/packages/toolchain-riscv32-esp/bin",
+            "internalConsoleOptions": "openOnSessionStart"
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (without uploading)",
+            "executable": "/Users/kevinlutzer/Projects/room-environment-monitor/rem-iot/.pio/build/esp32-c3-devkitm-1/firmware.elf",
+            "projectEnvName": "esp32-c3-devkitm-1",
+            "toolchainBinDir": "/Users/kevinlutzer/.platformio/packages/toolchain-riscv32-esp/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "loadMode": "manual"
+        }
+    ]
+}
