--- rem-iot/settings/settings.cpp
@@ -1,13 +1,0 @@
-#include "settings.hpp"
-
-Settings::Settings() {
-    
-}
-
-void Settings::serialize(uint8_t * buf, int len) {
-
-}
-
-void Settings::deserialize(uint8_t * buf, int len) {
-    
-}
